requirements.txt
Django==3.2.23
djangorestframework==3.12.0
djangorestframework-simplejwt==5.3.0
mysqlclient==2.2.0
pillow==10.4.0
python-dotenv==1.0.1
django-cors-headers==4.4.0
drf-yasg==1.21.10